Refactor error handling for external functions
The cause of an error is now only passed when calling external functions,
and is no longer included in the `RuntimeError` constructor. This simplifies
the API by removing unused fields in most cases.

This change is a precursor to updating the error interface to support error
codes in future commits.